optimized Enumerators for faster forach and LINQ
